+++ src/exchange_rate/mod.rs
@@ -14,8 +14,10 @@ use thiserror::Error;
 /// An error associated with functionality in the [exchange_rate](crate::exchange_rate) module.
 #[derive(Error, Debug)]
 pub enum ExchangeRateError {
-    #[error("the commodity type with id {0} is not present in the exchange rate")]
+    #[error("The commodity type with id {0} is not present in the exchange rate.")]
     CommodityTypeNotPresent(CommodityTypeID),
+    #[error("There was a divide overflow while computing the exchange rate, performing the division {0}/{1}.")]
+    DivideOverflow(Decimal, Decimal),
 }
 
 /// Represents the exchange rate between [Commodity](Commodity)s
@@ -61,10 +63,11 @@ impl ExchangeRate {
 
             if target_commodity_type == base {
                 if let Some(rate) = self.get_rate(&commodity.type_id) {
-                    return Ok(Commodity::new(
-                        commodity.value / rate,
-                        target_commodity_type,
-                    ));
+                    let div = commodity
+                        .value
+                        .checked_div(*rate)
+                        .ok_or_else(|| ExchangeRateError::DivideOverflow(commodity.value, *rate))?;
+                    return Ok(Commodity::new(div, target_commodity_type));
                 };
             }
         }
@@ -90,9 +93,57 @@ impl ExchangeRate {
             }
         };
 
-        let value = (commodity.value / commodity_rate) * target_rate;
+        let div = commodity
+            .value
+            .checked_div(*commodity_rate)
+            .ok_or_else(|| ExchangeRateError::DivideOverflow(commodity.value, *commodity_rate))?;
+        let value = div * target_rate;
 
         Ok(Commodity::new(value, target_commodity_type))
+    }
+
+    /// Get the exchange rate between two commodity types present in this exchange
+    /// rate data structure. Returns `None` if one of the commodity types is not present.
+    pub fn rate_between(
+        &self,
+        from: &CommodityTypeID,
+        to: &CommodityTypeID,
+    ) -> Result<Option<Decimal>, ExchangeRateError> {
+        if let Some(base) = &self.base {
+            if from == base {
+                if let Some(rate) = self.get_rate(&to) {
+                    return Ok(Some(*rate));
+                };
+            }
+
+            if to == base {
+                if let Some(rate) = self.get_rate(&from) {
+                    let one = Decimal::new(1, 0);
+                    return match one.checked_div(*rate) {
+                        Some(value) => Ok(Some(value)),
+                        None => Err(ExchangeRateError::DivideOverflow(one, *rate)),
+                    };
+                };
+            }
+        }
+
+        // handle the situation where there is no base commodity type, or neither the from
+        // or the to commodity types are the base commodity type.
+
+        let from_rate = match self.get_rate(&from) {
+            Some(rate) => rate,
+            None => return Ok(None),
+        };
+
+        let to_rate = match self.get_rate(&to) {
+            Some(rate) => rate,
+            None => return Ok(None),
+        };
+
+        match to_rate.checked_div(*from_rate) {
+            Some(value) => Ok(Some(value)),
+            None => Err(ExchangeRateError::DivideOverflow(*to_rate, *from_rate)),
+        }
     }
 }
 
@@ -151,6 +202,7 @@ mod tests {
         assert_eq!(expected_serialized_data, serialized_data);
     }
 
+    /// Convert between two commodities at a reference rate (no base rate commodity type).
     #[test]
     fn convert_reference_rates() {
         let mut rates: BTreeMap<CommodityTypeID, Decimal> = BTreeMap::new();
@@ -173,6 +225,10 @@ mod tests {
                 Decimal::from_str("10.412377413656575501005055735").unwrap(),
                 converted_commodity.unwrap().value
             );
+            assert_eq!(
+                exchange_rate.rate_between(&aud, &nzd).unwrap(),
+                Some(Decimal::from_str("1.0412377413656575501005055734").unwrap())
+            );
         }
 
         {
@@ -182,15 +238,23 @@ mod tests {
                 Decimal::from_str("9.603954603954603954603954604").unwrap(),
                 converted_commodity.unwrap().value
             );
+            assert_eq!(
+                exchange_rate.rate_between(&nzd, &aud).unwrap(),
+                Some(Decimal::from_str("0.9603954603954603954603954603").unwrap())
+            );
         }
     }
 
+    /// Convert between commodities using an exchange rate with a base rate.
     #[test]
     fn convert_base_rate() {
         let mut rates: BTreeMap<CommodityTypeID, Decimal> = BTreeMap::new();
         let nok = CommodityTypeID::from_str("NOK").unwrap();
         let usd = CommodityTypeID::from_str("USD").unwrap();
+        let gel = CommodityTypeID::from_str("GEL").unwrap();
+
         rates.insert(nok, Decimal::from_str("9.2691220713").unwrap());
+        rates.insert(gel, Decimal::from_str("3.08").unwrap());
 
         let exchange_rate = ExchangeRate {
             date: Some(NaiveDate::from_ymd(2020, 02, 07)),
@@ -206,6 +270,10 @@ mod tests {
                 Decimal::from_str("926.91220713000").unwrap(),
                 converted_commodity.unwrap().value
             );
+            assert_eq!(
+                exchange_rate.rate_between(&usd, &nok).unwrap(),
+                Some(Decimal::from_str("9.2691220713").unwrap())
+            );
         }
 
         {
@@ -214,6 +282,23 @@ mod tests {
             assert_eq!(
                 Decimal::from_str("10.788508256853169187585300627").unwrap(),
                 converted_commodity.unwrap().value
+            );
+            assert_eq!(
+                exchange_rate.rate_between(&nok, &usd).unwrap(),
+                Some(Decimal::from_str("0.1078850825685316918758530063").unwrap())
+            );
+        }
+
+        {
+            let start_commodity = Commodity::new(Decimal::from_str("100.0").unwrap(), nok);
+            let converted_commodity = exchange_rate.convert(start_commodity, gel);
+            assert_eq!(
+                Decimal::from_str("33.228605431107761097762725931").unwrap(),
+                converted_commodity.unwrap().value
+            );
+            assert_eq!(
+                exchange_rate.rate_between(&nok, &gel).unwrap(),
+                Some(Decimal::from_str("0.3322860543110776109776272593").unwrap())
             );
         }
     }